requirements.txt
fastapi==0.110.0
paho-mqtt<2.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests==2.22.0
SQLAlchemy==2.0.28
uvicorn==0.28.0
pytest==7.3.2